moved and renamed test
